requirements.txt
absl-py==2.2.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
batinfo==0.4.2
beautifulsoup4==4.13.4
bernhard==0.2.6
binaryornot==0.4.4
cassandra-driver==3.29.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.2
cookiecutter==2.6.0
cryptography==44.0.3
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.2
django-browser-reload==1.18.0
django-extensions==4.1
django-filebrowser-no-grappelli==4.0.2
django-import-export==4.3.7
django-multiselectfield==0.1.13
django-tailwind==4.0.1
djangorestframework==3.16.0
dnspython==2.7.0
docker==7.1.0
easydict==1.13
elastic-transport==8.17.1
elasticsearch==9.0.1
executing==2.2.0
fastapi==0.115.12
ffmpeg-python==0.2.0
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2024.6.1
future==1.0.0
gast==0.6.0
gdown==5.2.0
geomet==0.2.1.post1
git-filter-repo==2.47.0
Glances==4.3.1
google-pasta==0.2.0
graphitesender==0.11.2
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
humanfriendly==10.0
ibm-cloud-sdk-core==3.23.0
ibmcloudant==0.10.2
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
influxdb==5.3.2
influxdb-client==1.48.0
insightface==0.7.3
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
kafka-python==2.2.6
keras==3.9.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
netifaces2==0.0.22
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.3.0.75
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
onnx==1.17.0
onnxruntime-gpu==1.21.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
paho-mqtt==2.1.0
parso==0.8.4
pbkdf2==1.3
pexpect==4.9.0
pika==1.3.2
pillow==10.4.0
pip-check==3.1
ply==3.11
podman==5.4.0.1
potsdb==1.0.3
prettytable==3.16.0
proglog==0.1.11
prometheus_client==0.21.1
prompt_toolkit==3.0.51
protobuf==5.29.4
psutil==7.0.0
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pygal==3.0.5
Pygments==2.19.1
PyJWT==2.10.1
pymdstat==0.4.3
pymongo==4.12.1
pyparsing==3.2.3
pySMART.smartx==0.3.10
pysmi-lextudio==1.4.3
pysnmp-lextudio==6.1.2
pysnmpcrypto==0.0.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
reactivex==4.0.4
requests==2.32.3
retina-face==0.0.17
rich==14.0.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sparklines==0.5.0
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
statsd==4.0.1
stringzilla==3.12.4
sympy==1.13.1
tablib==3.8.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
terminaltables==3.1.10
text-unidecode==1.3
tf_keras==2.19.0
threadpoolctl==3.6.0
tifffile==2025.3.30
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
Werkzeug==3.1.3
wifi==0.3.8
wrapt==1.17.2
zeroconf==0.147.0
zipp==3.21.0
